// src/types/RequestInstitutionFromAdmin.tsx
    // private String email;
    // private String password;

    // private String businessRegistrationNumber;
    // private String institutionName;
    // private EducationInstitutionType institutionType;
    // private String phoneNumber;
    // private String website;
    // private String address;
    // private String representativeName;
    // private String position;
    // private String description;
    // private String logoUrl;
    // private Integer establishedYear;

export type RequestInstitutionFromAdmin = {
    email: string;
    password: string;
    
    businessRegistrationNumber: string;
    institutionName: string;
    institutionType: string;
    phoneNumber: string;
    website: string;
    address: string;
    representativeName: string;
    position: string;
    description: string;
    logoUrl: string;
    establishedYear: number;


}